reorganized ActiveGame components to better manage queries/subscriptions
