+++ src/pages/ActiveGame/PlayerPanel/index.js
@@ -5,18 +5,11 @@ import { Link } from "react-router-dom";
 import { Col, Row, Form, Image } from "react-bootstrap";
 import Container from "react-bootstrap/Container";
 
-import bugIcon from "../../images/icons/bugIcon.png";
-import eggIcon from "../../images/icons/eggIcon.png";
-import featherIcon from "../../images/icons/featherIcon.png";
-import marketIcon from "../../images/icons/marketIcon.png";
-import moneyCashIcon from "../../images/icons/moneyCashIcon.png";
-import rareIcon from "../../images/icons/rareIcon.png";
-import vPointIcon from "../../images/icons/vPointIcon.png";
-import { inlineIconStyle, iconStyle } from "../../styles/imgStyles";
-import PublicChat from "./PublicChat";
+import { inlineIconStyle, iconStyle } from "../../../styles/imgStyles";
+import PublicChat from "../PublicChat";
 import Player from "./Player";
 
-import Trade from "./SentTrade";
+import Trade from "../TradePanel/SentTrade";
 
 export default function PlayerPanel(props) {
   return (
