+++ src/pages/ActiveGame/PrivateChat/index.js
@@ -5,18 +5,18 @@ import { Link } from "react-router-dom";
 import { Col, Row, Form, Image, Alert } from "react-bootstrap";
 import Container from "react-bootstrap/Container";
 
-import bugIcon from "../../images/icons/bugIcon.png";
-import eggIcon from "../../images/icons/eggIcon.png";
-import featherIcon from "../../images/icons/featherIcon.png";
-import marketIcon from "../../images/icons/marketIcon.png";
-import moneyCashIcon from "../../images/icons/moneyCashIcon.png";
-import rareIcon from "../../images/icons/rareIcon.png";
-import vPointIcon from "../../images/icons/vPointIcon.png";
-import { inlineIconStyle, iconStyle } from "../../styles/imgStyles";
-import { GET_PRIVATE_MESSAGES_BY_ID } from "../../graphql/queries";
+// import bugIcon from "../../images/icons/bugIcon.png";
+// import eggIcon from "../../images/icons/eggIcon.png";
+// import featherIcon from "../../images/icons/featherIcon.png";
+// import marketIcon from "../../images/icons/marketIcon.png";
+// import moneyCashIcon from "../../images/icons/moneyCashIcon.png";
+// import rareIcon from "../../images/icons/rareIcon.png";
+// import vPointIcon from "../../images/icons/vPointIcon.png";
+// import { inlineIconStyle, iconStyle } from "../../styles/imgStyles";
+import { GET_PRIVATE_MESSAGES_BY_ID } from "../../../graphql/queries";
 import { useQuery, useSubscription } from "@apollo/react-hooks";
 
-import Trade from "./SentTrade";
+// import Trade from "./SentTrade";
 
 export default function PrivateChat(props) {
   const [inputField, set_inputField] = useState("hoi");
