+++ src/pages/ActiveGame/Scoreboard/index.js
@@ -2,16 +2,16 @@ import React, { useState, useEffect } from "react";
 import Form from "react-bootstrap/Form";
 import Container from "react-bootstrap/Container";
 import Button from "react-bootstrap/Button";
-import { login } from "../../store/user/actions";
-import { selectToken } from "../../store/user/selectors";
+import { login } from "../../../store/user/actions";
+import { selectToken } from "../../../store/user/selectors";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory, Link } from "react-router-dom";
 import { Col, Row } from "react-bootstrap";
 import PlayerScore from "./PlayerScore";
-import Player from "./Player";
-import TradePanel from "./TradePanel";
-import PublicChat from "./PublicChat";
-import PrivateChat from "./PrivateChat";
+import Player from "../PlayerPanel/Player";
+import TradePanel from "../TradePanel";
+import PublicChat from "../PublicChat";
+import PrivateChat from "../PrivateChat";
 import { useQuery, useSubscription } from "@apollo/react-hooks";
 
 export default function ScoreBoard(props) {
