+++ src/pages/ActiveGame/TradePanel/IncomingTrade.js
@@ -5,40 +5,40 @@ import { Link } from "react-router-dom";
 import { Col, Row, Image, Alert } from "react-bootstrap";
 import Container from "react-bootstrap/Container";
 import { useQuery, useSubscription, useMutation } from "@apollo/react-hooks";
-import { CLOSE_TRADE, ACCEPT_TRADE } from "../../graphql/mutations";
-import { GET_TRADES_BY_ID } from "../../graphql/queries";
+import { CLOSE_TRADE, ACCEPT_TRADE } from "../../../graphql/mutations";
+import { GET_TRADES_BY_ID } from "../../../graphql/queries";
 
-import bugIcon from "../../images/icons/bugIcon.png";
-import eggIcon from "../../images/icons/eggIcon.png";
-import featherIcon from "../../images/icons/featherIcon.png";
-import marketIcon from "../../images/icons/marketIcon.png";
-import moneyCashIcon from "../../images/icons/moneyCashIcon.png";
-import rareIcon from "../../images/icons/rareIcon.png";
-import vPointIcon from "../../images/icons/vPointIcon.png";
-import { inlineIconStyle, iconStyle } from "../../styles/imgStyles";
+import bugIcon from "../../../images/icons/bugIcon.png";
+import eggIcon from "../../../images/icons/eggIcon.png";
+import featherIcon from "../../../images/icons/featherIcon.png";
+import marketIcon from "../../../images/icons/marketIcon.png";
+import moneyCashIcon from "../../../images/icons/moneyCashIcon.png";
+import rareIcon from "../../../images/icons/rareIcon.png";
+import vPointIcon from "../../../images/icons/vPointIcon.png";
+import { inlineIconStyle, iconStyle } from "../../../styles/imgStyles";
 
-export default function SentTrade(props) {
+export default function IncomingTrade(props) {
+  // console.log("WIOUWGEWYGEUY", props.traderSenderId);
   const [closeTrade] = useMutation(CLOSE_TRADE);
   const [acceptTrade] = useMutation(ACCEPT_TRADE);
-  console.log("PROP", props.traderReceiverId);
   const { data, error, loading } = useQuery(GET_TRADES_BY_ID, {
     variables: {
-      playerSenderId: 1,
-      playerReceiverId: props.traderReceiverId,
+      playerSenderId: props.traderSenderId,
+      playerReceiverId: 1,
     },
   });
   if (loading) return "Loading...";
   if (error) return <Alert variant="danger">Error! {error.message}</Alert>;
-  // console.log("BONJOURR SENT", data);
+  // console.log("GOEIEDAG INCOMING", data);
 
-  if (data.getTradesById === null)
+  if (data.getTradesById === null) {
     return (
       <>
         <Card.Header>No open trade</Card.Header>
         <Card.Body>{""}</Card.Body>
       </>
     );
-  else {
+  } else {
     const {
       id,
       moneyCashSender,
@@ -50,6 +50,42 @@ export default function SentTrade(props) {
       bugSender,
       bugReceiver,
     } = data.getTradesById;
+
+    const tradeAccepter = (event) => {
+      // event.preventDefault();
+      acceptTrade({
+        variables: {
+          id,
+          playerSenderId: data.getTradesById.playerSenderId.id,
+          playerReceiverId: data.getTradesById.playerReceiverId.id,
+          moneyCashSender,
+          moneyCashReceiver,
+          eggSender,
+          eggReceiver,
+          featherSender,
+          featherReceiver,
+          bugSender,
+          bugReceiver,
+        },
+      });
+      window.location.reload(false);
+    };
+
+    // id
+    // playerSenderId {
+    //   id
+    // }
+    // playerReceiverId {
+    //   id
+    // }
+    // moneyCashSender
+    // moneyCashReceiver
+    // eggSender
+    // eggReceiver
+    // featherSender
+    // featherReceiver
+    // bugSender
+    // bugReceiver
 
     const noValueChecker = (resource, resourceIcon) => {
       if (resource != 0) {
@@ -69,12 +105,12 @@ export default function SentTrade(props) {
       <Card>
         {!data || !data.getTradesById.closed ? (
           <>
-            <Card.Header>Suggested Trade</Card.Header>
+            <Card.Header>Incoming trade</Card.Header>
             <Card.Body>
               <Card.Text as="div">
                 <Row>
                   <Col>
-                    <Row>{data.getTradesById.playerSenderId.name} offers: </Row>
+                    {data.getTradesById.playerSenderId.name} offers: <br></br>
                     {noValueChecker(moneyCashSender, moneyCashIcon)}
                     {noValueChecker(eggSender, eggIcon)}
                     {noValueChecker(featherSender, featherIcon)}
@@ -95,10 +131,19 @@ export default function SentTrade(props) {
               <Row>
                 <Col>
                   <Button
+                    variant="outline-success"
+                    size="sm"
+                    onClick={tradeAccepter}
+                  >
+                    Accept
+                  </Button>
+                </Col>
+                <Col>
+                  <Button
                     variant="outline-danger"
                     size="sm"
                     onClick={(event) => {
-                      console.log("Cancel trade");
+                      console.log("Decline trade");
                       // event.preventDefault();
                       closeTrade({
                         variables: { id, closed: true },
@@ -106,7 +151,7 @@ export default function SentTrade(props) {
                       window.location.reload(false);
                     }}
                   >
-                    Cancel
+                    Decline
                   </Button>
                 </Col>
               </Row>
