+++ src/pages/ActiveGame/TradePanel/SentTrade.js
@@ -5,40 +5,40 @@ import { Link } from "react-router-dom";
 import { Col, Row, Image, Alert } from "react-bootstrap";
 import Container from "react-bootstrap/Container";
 import { useQuery, useSubscription, useMutation } from "@apollo/react-hooks";
-import { CLOSE_TRADE, ACCEPT_TRADE } from "../../graphql/mutations";
-import { GET_TRADES_BY_ID } from "../../graphql/queries";
+import { CLOSE_TRADE, ACCEPT_TRADE } from "../../../graphql/mutations";
+import { GET_TRADES_BY_ID } from "../../../graphql/queries";
 
-import bugIcon from "../../images/icons/bugIcon.png";
-import eggIcon from "../../images/icons/eggIcon.png";
-import featherIcon from "../../images/icons/featherIcon.png";
-import marketIcon from "../../images/icons/marketIcon.png";
-import moneyCashIcon from "../../images/icons/moneyCashIcon.png";
-import rareIcon from "../../images/icons/rareIcon.png";
-import vPointIcon from "../../images/icons/vPointIcon.png";
-import { inlineIconStyle, iconStyle } from "../../styles/imgStyles";
+import bugIcon from "../../../images/icons/bugIcon.png";
+import eggIcon from "../../../images/icons/eggIcon.png";
+import featherIcon from "../../../images/icons/featherIcon.png";
+import marketIcon from "../../../images/icons/marketIcon.png";
+import moneyCashIcon from "../../../images/icons/moneyCashIcon.png";
+import rareIcon from "../../../images/icons/rareIcon.png";
+import vPointIcon from "../../../images/icons/vPointIcon.png";
+import { inlineIconStyle, iconStyle } from "../../../styles/imgStyles";
 
-export default function IncomingTrade(props) {
-  // console.log("WIOUWGEWYGEUY", props.traderSenderId);
+export default function SentTrade(props) {
   const [closeTrade] = useMutation(CLOSE_TRADE);
   const [acceptTrade] = useMutation(ACCEPT_TRADE);
+  console.log("PROP", props.traderReceiverId);
   const { data, error, loading } = useQuery(GET_TRADES_BY_ID, {
     variables: {
-      playerSenderId: props.traderSenderId,
-      playerReceiverId: 1,
+      playerSenderId: 1,
+      playerReceiverId: props.traderReceiverId,
     },
   });
   if (loading) return "Loading...";
   if (error) return <Alert variant="danger">Error! {error.message}</Alert>;
-  // console.log("GOEIEDAG INCOMING", data);
+  // console.log("BONJOURR SENT", data);
 
-  if (data.getTradesById === null) {
+  if (data.getTradesById === null)
     return (
       <>
         <Card.Header>No open trade</Card.Header>
         <Card.Body>{""}</Card.Body>
       </>
     );
-  } else {
+  else {
     const {
       id,
       moneyCashSender,
@@ -50,42 +50,6 @@ export default function IncomingTrade(props) {
       bugSender,
       bugReceiver,
     } = data.getTradesById;
-
-    const tradeAccepter = (event) => {
-      // event.preventDefault();
-      acceptTrade({
-        variables: {
-          id,
-          playerSenderId: data.getTradesById.playerSenderId.id,
-          playerReceiverId: data.getTradesById.playerReceiverId.id,
-          moneyCashSender,
-          moneyCashReceiver,
-          eggSender,
-          eggReceiver,
-          featherSender,
-          featherReceiver,
-          bugSender,
-          bugReceiver,
-        },
-      });
-      window.location.reload(false);
-    };
-
-    // id
-    // playerSenderId {
-    //   id
-    // }
-    // playerReceiverId {
-    //   id
-    // }
-    // moneyCashSender
-    // moneyCashReceiver
-    // eggSender
-    // eggReceiver
-    // featherSender
-    // featherReceiver
-    // bugSender
-    // bugReceiver
 
     const noValueChecker = (resource, resourceIcon) => {
       if (resource != 0) {
@@ -105,12 +69,12 @@ export default function IncomingTrade(props) {
       <Card>
         {!data || !data.getTradesById.closed ? (
           <>
-            <Card.Header>Incoming trade</Card.Header>
+            <Card.Header>Suggested Trade</Card.Header>
             <Card.Body>
               <Card.Text as="div">
                 <Row>
                   <Col>
-                    {data.getTradesById.playerSenderId.name} offers: <br></br>
+                    <Row>{data.getTradesById.playerSenderId.name} offers: </Row>
                     {noValueChecker(moneyCashSender, moneyCashIcon)}
                     {noValueChecker(eggSender, eggIcon)}
                     {noValueChecker(featherSender, featherIcon)}
@@ -131,19 +95,10 @@ export default function IncomingTrade(props) {
               <Row>
                 <Col>
                   <Button
-                    variant="outline-success"
-                    size="sm"
-                    onClick={tradeAccepter}
-                  >
-                    Accept
-                  </Button>
-                </Col>
-                <Col>
-                  <Button
                     variant="outline-danger"
                     size="sm"
                     onClick={(event) => {
-                      console.log("Decline trade");
+                      console.log("Cancel trade");
                       // event.preventDefault();
                       closeTrade({
                         variables: { id, closed: true },
@@ -151,7 +106,7 @@ export default function IncomingTrade(props) {
                       window.location.reload(false);
                     }}
                   >
-                    Decline
+                    Cancel
                   </Button>
                 </Col>
               </Row>
