+++ src/pages/ActiveGame/TradePanel/index.js
@@ -6,23 +6,16 @@ import { Col, Row, Form, Image, Alert } from "react-bootstrap";
 import Container from "react-bootstrap/Container";
 import { useDispatch, useSelector } from "react-redux";
 
-import bugIcon from "../../images/icons/bugIcon.png";
-import eggIcon from "../../images/icons/eggIcon.png";
-import featherIcon from "../../images/icons/featherIcon.png";
-import marketIcon from "../../images/icons/marketIcon.png";
-import moneyCashIcon from "../../images/icons/moneyCashIcon.png";
-import rareIcon from "../../images/icons/rareIcon.png";
-import vPointIcon from "../../images/icons/vPointIcon.png";
-import { inlineIconStyle, iconStyle } from "../../styles/imgStyles";
+import { inlineIconStyle, iconStyle } from "../../../styles/imgStyles";
 import { useQuery, useSubscription } from "@apollo/react-hooks";
-import { GET_TRADES_BY_ID } from "../../graphql/queries";
-import { storeTradePlayer } from "../../store/tradePlayer/actions";
+import { GET_TRADES_BY_ID } from "../../../graphql/queries";
+import { storeTradePlayer } from "../../../store/tradePlayer/actions";
 
-import PrivateChat from "./PrivateChat";
+import PrivateChat from "../PrivateChat";
 import SentTrade from "./SentTrade";
 import IncomingTrade from "./IncomingTrade";
 import TradeSuggest from "./TradeSuggest";
-import PublicChat from "./PublicChat";
+import PublicChat from "../PublicChat";
 
 export default function TradePanel(props) {
   const dispatch = useDispatch();
